chore: Move last query over
